api/requirements.txt
Flask==3.0.0
fpdf2==2.7.8
flask-cors==4.0.0
